requirements.txt
streamlit
requests
beautifulsoup4
lxml
selenium
webdriver-manager
google-generativeai
